Move TestRun to separate file
